setting new remote
