+++ models/models.js
@@ -1,24 +1,37 @@
-const { fetchTopics, fetchArticles, fetchArticleById } = require('../controllers/controllers.js')
+const format = require('pg-format');
+const db = require('../db/connection');
 
-const getTopics = (req, res) => { 
-    fetchTopics().then(data => {
-        res.status(200).send(data); 
+const fetchTopics = () => { 
+    return db.query('SELECT * FROM topics;')
+    .then(topics => {
+        return topics.rows;
+    })
+    .catch(err => {
+        console.log(err);
     });
 }
 
-const getArticles = (req, res) => { 
-    fetchArticles().then(data => {
-        return res.status(200).send(data);
+const fetchArticles = () => { 
+    return db.query('SELECT * FROM articles;')
+    .then(topics => {
+        return topics.rows;
+    })
+    .catch(err => {
+        console.log(err);
     });
 }
 
-const getArticleById = (req, res) => {
-    const id = Number(req.params.id);
-    fetchArticleById(id).then(data => {
-        return res.status(200).send(data);
+const fetchArticleById = (id) => { 
+    id = [id];
+    return db.query('SELECT * FROM articles WHERE article_id = $1;', id)
+    .then(topics => {
+        return topics.rows;
+    })
+    .catch(err => {
+        console.log(err);
     });
 }
 
 module.exports = {
-    getTopics, getArticles, getArticleById
+    fetchTopics, fetchArticles, fetchArticleById
 }
